createStudyBrowserTabsUtil
