+++ platform/core/src/utils/index.ts
@@ -37,6 +37,7 @@ import {
 } from './sortStudy';
 import { subscribeToNextViewportGridChange } from './subscribeToNextViewportGridChange';
 import { splitComma, getSplitParam } from './splitComma';
+import { createStudyBrowserTabs } from './createStudyBrowserTabs';
 
 // Commented out unused functionality.
 // Need to implement new mechanism for derived displaySets using the displaySetManager.
@@ -80,6 +81,7 @@ const utils = {
   splitComma,
   getSplitParam,
   generateAcceptHeader,
+  createStudyBrowserTabs,
 };
 
 export {
@@ -112,6 +114,7 @@ export {
   splitComma,
   getSplitParam,
   generateAcceptHeader,
+  createStudyBrowserTabs,
 };
 
 export default utils;
